Added ability to read geometry
Implemented ability to read geometry from a config file - this is limited to reading in circles at the moment.

A new class has been generated which aims to hold information about each of the bodies which are generated - this is: ```IBMBodyClass()```.

In order to add more functionality the generated bodies will need to produce *nodes* in order to carry out the required calculation.

Resolves: #50