Implemented computeEpsilon()
Implemented ```ObjectsClass::computeEpsilon()``` - this allows for succesful initialisation of objects within the grid. In order to complete this additional code was required for solving a system of equations - this can be found within: ```Utils::solveLAPACK()``` all of the external calls are found within the ```<Accelerate/Accelerate.h> header.

The ```README.md``` has been updated to reflect the change in terminal command needed to compile the code.

Resolves: #65